division of controllers and services and reservation controllers reservation-detail.html styling

+++ reserve_point/src/main/java/cz/cervenka/reserve_point/database/repositories/ReservationRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 import java.time.LocalDate;
 import java.util.List;
 
-public interface ReservationRepository extends JpaRepository<ReservationEntity, Integer> {
+public interface ReservationRepository extends JpaRepository<ReservationEntity, Long> {
 
     @Query("SELECT r FROM ReservationEntity r " +
             "JOIN r.customer c " +
